update backend styles and log/index

+++ backend/views/log/index.php
@@ -2,6 +2,9 @@
 
 use yii\helpers\Html;
 use yii\grid\GridView;
+use yii\widgets\Pjax;
+use yii\helpers\Url;
+use backend\models\SystemLog;
 
 /* @var $this yii\web\View */
 /* @var $searchModel backend\models\search\SystemLogSearch */
@@ -13,11 +16,20 @@ $this->params['breadcrumbs'][] = $this->title;
 <div class="system-log-index">
 
     <p>
-        <?= Html::a(Yii::t('backend', 'Clear'), false, ['class' => 'btn btn-danger', 'data-method'=>'delete']) ?>
+        <?= Html::a(Yii::t('backend', 'Clear'), false, ['class' => 'btn btn-danger', 'data-method'=>'delete', 'data-confirm' => Yii::t('yii', 'Are you sure you want to delete this item?')]) ?>
     </p>
+    <?php $ranges = [
+        Yii::t('kvdrp', "Today") => ["moment().startOf('day')", "moment().startOf('day').add({days:1})"],
+        Yii::t('kvdrp', "Yesterday") => ["moment().startOf('day').subtract(1,'days')", "moment().startOf('day')"],
+        Yii::t('kvdrp', "Last {n} Days", ['n' => 7]) => ["moment().startOf('day').subtract(6, 'days')", "moment().startOf('day').add({days:1})"],
+        Yii::t('kvdrp', "Last {n} Days", ['n' => 14]) => ["moment().startOf('day').subtract(13, 'days')", "moment().startOf('day').add({days:1})"],
+        Yii::t('kvdrp', "This Month") => ["moment().startOf('month')", "moment().startOf('month').add({month:1})"],
+        Yii::t('kvdrp', "Last Month") => ["moment().subtract({month:1}).startOf('day')", "moment().startOf('day').add({days:1})"],
+    ]; ?>
 
     <?//= $this->render('_search', ['model' => $searchModel]); ?>
 
+    <?php Pjax::begin(); ?>
     <?= GridView::widget([
         'dataProvider' => $dataProvider,
         'filterModel' => $searchModel,
@@ -26,6 +38,18 @@ $this->params['breadcrumbs'][] = $this->title;
         ],
         'columns' => [
             ['class' => 'yii\grid\SerialColumn'],
+
+            [
+                'attribute' => 'prefix',
+                'format' => 'raw',
+                'value' => function (SystemLog $model) {
+                    return Html::a($model->prefix, Url::to(['view', 'id' => $model->id]));
+                },
+            ],
+            [
+                'attribute' => 'category',
+                'filter'=>SystemLog::find()->select(['category'])->indexBy('category')->distinct()->asArray()->column()
+            ],
             [
                 'attribute'=>'level',
                 'value'=>function ($model) {
@@ -40,21 +64,29 @@ $this->params['breadcrumbs'][] = $this->title;
                     \yii\log\Logger::LEVEL_PROFILE_END => 'profile end'
                 ]
             ],
-            'category',
-            'prefix',
             [
                 'attribute' => 'log_time',
+                'value' => 'log_time',
                 'format' => 'datetime',
-                'value' => function ($model) {
-                    return (int) $model->log_time;
-                }
+                'filter'=>\kartik\daterange\DateRangePicker::widget([
+                    'model' => $searchModel,
+                    'attribute' => 'log_time',
+                    'convertFormat' => true,
+                    'pluginOptions' => [
+                        'ranges' => $ranges,
+                        'locale' => [
+                            'format' => 'Y-m-d',
+                        ],
+                    ],
+                ]),
             ],
 
             [
                 'class' => 'yii\grid\ActionColumn',
-                'template'=>'{view}{delete}'
+                'template'=>'{delete}'//{view}
             ]
         ]
     ]); ?>
+    <?php Pjax::end(); ?>
 
 </div>
